# backend-order/docs/swagger.yaml
basePath: /
definitions:
...
  api.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  api.LoginResponse:
    properties:
      token:
        type: string
    type: object
  api.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  api.ResetPasswordRequest:
    properties:
      email:
        type: string
      newPassword:
        minLength: 6
        type: string
      resetToken:
        type: string
    required:
    - email
    - newPassword
    - resetToken
    type: object
  backend.PaymentUpdateRequest:
    properties:
      amount:
        type: number
      order_id:
        type: string
      status:
        type: string
    required:
    - amount
    - order_id
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a simple backend server using Go and Gin framework.
  title: Order API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset token to the user's email
      parameters:
      - description: User's email
        in: body
        name: forgotRequest
        required: true
        schema:
          $ref: '#/definitions/api.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate forgot password process
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration details
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/api.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset a user's password
      parameters:
      - description: Password reset details
        in: body
        name: resetRequest
        required: true
        schema:
          $ref: '#/definitions/api.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset user password
      tags:
      - Authentication
  /backend/payment-update:
    post:
      consumes:
      - application/json
      description: Update the payment status of an order (backend communication)
      parameters:
      - description: Payment update details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/backend.PaymentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update order payment status
      tags:
      - Backend
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"